Added JSON data connection to mock API tests, in case no PG connection.

+++ MockAPILab/routes/ltk/index.js
@@ -2,10 +2,26 @@ const express = require('express');
 const router = express.Router();
 const pool = require('../../pool');
 const pgFormat = require('pg-format');
+const { jsonDefaultData, deleteBorrowerDummyData } = require('../../dummyData');
+let usePostgresDatabase = true;
+let useJsonData = false;
+
+
+// ⬇ Checking if the user has a Postgres Database connection established.  If not, we'll default to using the JSON data instead.
+pool.connect((err, client, release) => {
+	if (err) {
+		usePostgresDatabase = false;
+		useJsonData = true;
+		console.log('Error connecting to Postgres Database.  Using JSON data instead.', { usePostgresDatabase, useJsonData });
+	}; // End if
+	release();
+});
 
 
 // ⬇ Gets all of the Loan Data available in public.loans table:
 router.get('/fetch-all-loans', async (req, res) => {
+	if (useJsonData) return res.json(jsonDefaultData);
+
 	const sql = `SELECT * FROM loans`;
 
 	try {
@@ -21,6 +37,9 @@ router.get('/fetch-all-loans', async (req, res) => {
 // ⬇ Gets one loan object based off of the loanId:
 router.get('/fetch-loan-by-id/:id', async (req, res) => {
 	const loanId = req.params.id;
+
+	if (useJsonData) return res.json(jsonDefaultData.find(loan => loan.loanId === Number(loanId)));
+
 	const sql = `SELECT * FROM loans WHERE loanId = ${pgFormat('%L::int', loanId)}`;
 
 	try {
@@ -36,6 +55,13 @@ router.get('/fetch-loan-by-id/:id', async (req, res) => {
 // ⬇ Adds a new loan object to the public.loans table:
 router.post('/add-loan', async (req, res) => {
 	const newLoan = req.body;
+
+	if (useJsonData) {
+		newLoan.loanId = jsonDefaultData.length + 2;
+		jsonDefaultData.push(newLoan);
+		return res.sendStatus(201);
+	}; // End if
+
 	const sql = `INSERT INTO loans(borrowers) VALUES (${pgFormat('%L', newLoan)})`;
 
 	try {
@@ -52,6 +78,14 @@ router.post('/add-loan', async (req, res) => {
 router.put('/update-loan/:loanId/:paidId', async (req, res) => {
 	const { loanId, paidId } = req.params;
 	const updatedLoan = req.body;
+
+	if (useJsonData) {
+		const loan = jsonDefaultData.find(loan => loan.loanId === Number(loanId));
+		const borrower = loan.borrowers.find(borrower => borrower.paidId === Number(paidId));
+		borrower.firstName = updatedLoan.firstName;
+		return res.sendStatus(201);
+	}; // End if
+
 	updatedLoan.paidId = Math.floor(Math.random() * 1000000);
 
 	const sql = `
@@ -75,6 +109,17 @@ router.put('/update-loan/:loanId/:paidId', async (req, res) => {
 // ⬇ Deletes a borrower from a loan object in the public.loans table:
 router.put('/update-loan-delete-borrower/:loanId/:paidId', async (req, res) => {
 	const { loanId, paidId } = req.params;
+
+	if (useJsonData) {
+		const loan = jsonDefaultData.find(loan => loan.loanId === Number(loanId));
+		loan.borrowers = loan.borrowers.filter(borrower => borrower.paidId !== Number(paidId));
+
+		// ⬇ Adding the same user back-in to not break the front-end every test:
+		jsonDefaultData.push(deleteBorrowerDummyData);
+
+		return res.sendStatus(201);
+	}; // End if
+
 	const sql = `
 		UPDATE loans
 		SET borrowers = (
@@ -103,6 +148,18 @@ router.put('/update-loan-delete-borrower/:loanId/:paidId', async (req, res) => {
 // ⬇ Deletes a loan object from the public.loans table based on LoanId:
 router.delete('/delete-loan/:id', async (req, res) => {
 	const loanId = req.params.id;
+
+
+	if (useJsonData) {
+		jsonDefaultData.filter(loan => loan.loanId !== Number(loanId));
+
+		// ⬇ Adding the same user back-in to not break the front-end every test:
+		jsonDefaultData.push(deleteBorrowerDummyData);
+
+		return res.sendStatus(201);
+	}; // End if
+
+
 	const sql = `DELETE FROM loans WHERE loanId = ${pgFormat('%L::int', loanId)}`;
 
 	try {
